requirements.txt
fastapi
pinecone
requests
uvicorn
python-dotenv